refactor : 패키지 domain 별로 분리
