+++ src/main/java/cart/JwpCartApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class JwpCartApplication {
 
-    public static void main(String[] args) {
-        SpringApplication.run(JwpCartApplication.class, args);
-    }
+  public static void main(String[] args) {
+    SpringApplication.run(JwpCartApplication.class, args);
+  }
 
 }
