+++ src/main/java/cart/cart_item/application/dto/CartItemQuantityUpdateRequest.java
@@ -1,0 +1,17 @@
+package cart.cart_item.application.dto;
+
+public class CartItemQuantityUpdateRequest {
+
+  private int quantity;
+
+  public CartItemQuantityUpdateRequest() {
+  }
+
+  public CartItemQuantityUpdateRequest(int quantity) {
+    this.quantity = quantity;
+  }
+
+  public int getQuantity() {
+    return quantity;
+  }
+}
