+++ src/main/java/cart/cart_item/application/dto/CartItemRequest.java
@@ -1,0 +1,17 @@
+package cart.cart_item.application.dto;
+
+public class CartItemRequest {
+
+  private Long productId;
+
+  public CartItemRequest() {
+  }
+
+  public CartItemRequest(Long productId) {
+    this.productId = productId;
+  }
+
+  public Long getProductId() {
+    return productId;
+  }
+}
