+++ src/main/java/cart/cart_item/application/dto/CartItemResponse.java
@@ -1,0 +1,37 @@
+package cart.cart_item.application.dto;
+
+import cart.cart_item.domain.CartItem;
+import cart.product.application.dto.ProductResponse;
+
+public class CartItemResponse {
+
+  private Long id;
+  private int quantity;
+  private ProductResponse product;
+
+  private CartItemResponse(Long id, int quantity, ProductResponse product) {
+    this.id = id;
+    this.quantity = quantity;
+    this.product = product;
+  }
+
+  public static CartItemResponse of(CartItem cartItem) {
+    return new CartItemResponse(
+        cartItem.getId(),
+        cartItem.getQuantity(),
+        ProductResponse.of(cartItem.getProduct())
+    );
+  }
+
+  public Long getId() {
+    return id;
+  }
+
+  public int getQuantity() {
+    return quantity;
+  }
+
+  public ProductResponse getProduct() {
+    return product;
+  }
+}
