+++ src/main/java/cart/config/WebMvcConfig.java
@@ -1,0 +1,23 @@
+package cart.config;
+
+import cart.member.controller.MemberArgumentResolver;
+import cart.member.dao.MemberDao;
+import java.util.List;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+
+  private final MemberDao memberDao;
+
+  public WebMvcConfig(MemberDao memberDao) {
+    this.memberDao = memberDao;
+  }
+
+  @Override
+  public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+    resolvers.add(new MemberArgumentResolver(memberDao));
+  }
+}
