--- src/main/java/cart/domain/Member.java
@@ -1,29 +1,0 @@
-package cart.domain;
-
-public class Member {
-    private Long id;
-    private String email;
-    private String password;
-
-    public Member(Long id, String email, String password) {
-        this.id = id;
-        this.email = email;
-        this.password = password;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public boolean checkPassword(String password) {
-        return this.password.equals(password);
-    }
-}
