--- src/main/java/cart/dto/ProductRequest.java
@@ -1,28 +1,0 @@
-package cart.dto;
-
-public class ProductRequest {
-    private String name;
-    private int price;
-    private String imageUrl;
-
-    public ProductRequest() {
-    }
-
-    public ProductRequest(String name, int price, String imageUrl) {
-        this.name = name;
-        this.price = price;
-        this.imageUrl = imageUrl;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
-    }
-}
