+++ src/main/java/cart/exception/ControllerExceptionHandler.java
@@ -1,0 +1,22 @@
+package cart.exception;
+
+import cart.cart_item.exception.CartItemException;
+import cart.member.exception.AuthenticationException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class ControllerExceptionHandler {
+
+  @ExceptionHandler(AuthenticationException.class)
+  public ResponseEntity<Void> handlerAuthenticationException(AuthenticationException e) {
+    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+  }
+
+  @ExceptionHandler(CartItemException.IllegalMember.class)
+  public ResponseEntity<Void> handleException(CartItemException.IllegalMember e) {
+    return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+  }
+}
