+++ src/main/java/cart/member/exception/AuthenticationException.java
@@ -1,0 +1,5 @@
+package cart.member.exception;
+
+public class AuthenticationException extends RuntimeException {
+
+}
