+++ src/main/java/cart/product/controller/ProductApiController.java
@@ -1,0 +1,58 @@
+package cart.product.controller;
+
+import cart.product.application.ProductService;
+import cart.product.application.dto.ProductRequest;
+import cart.product.application.dto.ProductResponse;
+import java.net.URI;
+import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/products")
+public class ProductApiController {
+
+  private final ProductService productService;
+
+  public ProductApiController(ProductService productService) {
+    this.productService = productService;
+  }
+
+  @GetMapping
+  public ResponseEntity<List<ProductResponse>> getAllProducts() {
+    return ResponseEntity.ok(productService.getAllProducts());
+  }
+
+  @GetMapping("/{id}")
+  public ResponseEntity<ProductResponse> getProductById(@PathVariable Long id) {
+    return ResponseEntity.ok(productService.getProductById(id));
+  }
+
+  @PostMapping
+  public ResponseEntity<Void> createProduct(@RequestBody ProductRequest productRequest) {
+    Long id = productService.createProduct(productRequest);
+    return ResponseEntity.created(URI.create("/products/" + id)).build();
+  }
+
+  @PutMapping("/{id}")
+  public ResponseEntity<Void> updateProduct(@PathVariable Long id,
+      @RequestBody ProductRequest productRequest) {
+    productService.updateProduct(id, productRequest);
+    return ResponseEntity.status(HttpStatus.OK).build();
+  }
+
+  @DeleteMapping("/{id}")
+  public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
+    productService.deleteProduct(id);
+    return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+  }
+
+}
