--- src/main/java/cart/ui/CartItemApiController.java
@@ -1,50 +1,0 @@
-package cart.ui;
-
-import cart.application.CartItemService;
-import cart.domain.CartItem;
-import cart.domain.Member;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemRequest;
-import cart.dto.CartItemResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/cart-items")
-public class CartItemApiController {
-
-    private final CartItemService cartItemService;
-
-    public CartItemApiController(CartItemService cartItemService) {
-        this.cartItemService = cartItemService;
-    }
-
-    @GetMapping
-    public ResponseEntity<List<CartItemResponse>> showCartItems(Member member) {
-        return ResponseEntity.ok(cartItemService.findByMember(member));
-    }
-
-    @PostMapping
-    public ResponseEntity<Void> addCartItems(Member member, @RequestBody CartItemRequest cartItemRequest) {
-        Long cartItemId = cartItemService.add(member, cartItemRequest);
-
-        return ResponseEntity.created(URI.create("/cart-items/" + cartItemId)).build();
-    }
-
-    @PatchMapping("/{id}")
-    public ResponseEntity<Void> updateCartItemQuantity(Member member, @PathVariable Long id, @RequestBody CartItemQuantityUpdateRequest request) {
-        cartItemService.updateQuantity(member, id, request);
-
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> removeCartItems(Member member, @PathVariable Long id) {
-        cartItemService.remove(member, id);
-
-        return ResponseEntity.noContent().build();
-    }
-}
