+++ src/main/java/cart/ui/PageController.java
@@ -1,30 +1,31 @@
 package cart.ui;
 
-import cart.application.ProductService;
-import cart.dao.MemberDao;
+import cart.member.dao.MemberDao;
+import cart.product.application.ProductService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class PageController {
-    private final ProductService productService;
-    private final MemberDao memberDao;
 
-    public PageController(ProductService productService, MemberDao memberDao) {
-        this.productService = productService;
-        this.memberDao = memberDao;
-    }
+  private final ProductService productService;
+  private final MemberDao memberDao;
 
-    @GetMapping("/admin")
-    public String admin(Model model) {
-        model.addAttribute("products", productService.getAllProducts());
-        return "admin";
-    }
+  public PageController(ProductService productService, MemberDao memberDao) {
+    this.productService = productService;
+    this.memberDao = memberDao;
+  }
 
-    @GetMapping("/settings")
-    public String members(Model model) {
-        model.addAttribute("members", memberDao.getAllMembers());
-        return "settings";
-    }
+  @GetMapping("/admin")
+  public String admin(Model model) {
+    model.addAttribute("products", productService.getAllProducts());
+    return "admin";
+  }
+
+  @GetMapping("/settings")
+  public String members(Model model) {
+    model.addAttribute("members", memberDao.getAllMembers());
+    return "settings";
+  }
 }
