+++ src/test/java/cart/product/controller/ProductApiControllerIntegrationTest.java
@@ -1,0 +1,71 @@
+package cart.product.controller;
+
+import cart.helper.IntegrationTestHelper;
+import cart.product.application.dto.ProductRequest;
+import cart.product.application.dto.ProductResponse;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import static io.restassured.RestAssured.given;
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class ProductApiControllerIntegrationTest extends IntegrationTestHelper {
+
+    @Test
+    public void getProducts() {
+        var result = given()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when()
+                .get("/products")
+                .then()
+                .extract();
+
+        assertThat(result.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    @Test
+    public void createProduct() {
+        var product = new ProductRequest("치킨", 10_000, "http://example.com/chicken.jpg");
+
+        var response = given()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(product)
+                .when()
+                .post("/products")
+                .then()
+                .extract();
+
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+    }
+
+    @Test
+    public void getCreatedProduct() {
+        var product = new ProductRequest("피자", 15_000, "http://example.com/pizza.jpg");
+
+        // create product
+        var location =
+                given()
+                        .contentType(MediaType.APPLICATION_JSON_VALUE)
+                        .body(product)
+                        .when()
+                        .post("/products")
+                        .then()
+                        .statusCode(HttpStatus.CREATED.value())
+                        .extract().header("Location");
+
+        // get product
+        var responseProduct = given().log().all()
+                .when()
+                .get(location)
+                .then()
+                .statusCode(HttpStatus.OK.value())
+                .extract()
+                .jsonPath()
+                .getObject(".", ProductResponse.class);
+
+        assertThat(responseProduct.getId()).isNotNull();
+        assertThat(responseProduct.getName()).isEqualTo("피자");
+        assertThat(responseProduct.getPrice()).isEqualTo(15_000);
+    }
+}
